Doctrine types
