+++ src/Infrastructure/Persistence/Doctrine/Types/PostStatusType.php
@@ -4,7 +4,6 @@ namespace SocialNetworksPublisher\Infrastructure\Persistence\Doctrine\Types;
 
 use Phariscope\DoctrineEnumType\EnumType;
 use SocialNetworksPublisher\Domain\Model\Page\PostStatus;
-use SocialNetworksPublisher\Domain\Model\Shared\SocialNetworks;
 
 class PostStatusType extends EnumType
 {
